requirements.txt
discord.py
python.dotenv
discord==2.3.2
flask==3.1.2